% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_install.R
\name{safe_install_github}
\alias{safe_install_github}
\title{Safe installation of R packages from GitHub}
\usage{
safe_install_github(user.repo = NULL, ..., verbose = TRUE)
}
\arguments{
\item{user.repo}{Repository address in the format \code{username/repo}.}

\item{...}{Further options passed to \code{\link[remotes:install_github]{remotes::install_github()}}.}

\item{verbose}{Logical. Show verbose output?}
}
\value{
If no unsafe code is detected, the package is installed. Otherwise, an error message calling to inspect the repository for potential problems.
}
\description{
Safe installation of R packages from GitHub
}
\examples{
library(safeinstall)
safe_install_github("Pakillo/safeinstall")
safe_install_github("ropenscilabs/testevil")
}
