% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeinstall-package.R
\docType{package}
\name{safeinstall-package}
\alias{safeinstall}
\alias{safeinstall-package}
\title{safeinstall: Run some basic security checks before installing R packages from online repositories}
\description{
Run some basic security checks before installing R packages from online repositories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Pakillo/safeinstall}
  \item Report bugs at \url{https://github.com/Pakillo/safeinstall/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rodriguez-Sanchez \email{f.rodriguez.sanc@gmail.com}

}
\keyword{internal}
